Do not use project ID from jira
The jira project ID is only useful in the context of JIRA to retrieve
more metadata for that project. It is not meaningful when used
externally and does not match with the project table in the database.
Instead use the project ID already known from the project import (there
was already an assumption in import order anyway).