Fix status and resolution checks
The checks should use the name instead of the ID.
This caused duplicate information in the status database from each project.
The resolutions were not affected because it attempts to insert the correct ID.
Note that the id columns are set to be auto-increment even though they may not be.
All of this does not affect the issue data itself which uses the original JIRA identifiers.
These tables mainly serve a metadata/documentation purpose by providing a mapping to readable names.