Importer: Refactor base classes and methods, tasks
This extracts the basic functionality that both the importer and DAO DB
classes use to a BaseDb, so that the DBs are no longer importers. The
BaseImport now has both project ID and name, which are no longer passed
to every parse method. This makes it possible to refactor import tasks
process in the main class. We can now specify via arguments what we want
to import and skip certain tasks without changing the code/building.

Clean up some unused stuff and add more final markers.
Ensure epics are set to null instead of "0" in the database.