Fix git repo import and name hashing
The check for adding new repositories was wrong,
causing repos to never get added to the database.
Repository information is mostly metadata,
but the repo ID of each commit was also missing.

Additionally, hashing developer names would overwrite
the assignee and updater of an issue change with the reporter,
and hashing the comment author name fails due to table mismatch.

Clean up the commit importer a little bit.