Metrics: Fix name map not being filled
The metrics name map cache was filled with null values on a clean import
because the wrong value was used, thus a newly added metric name had to
be requested for each metric value with that name. Thus in this case the
name cache did not provide a performance improvement.

Also remove unused imports.