Importer: Use try-with-resources for closing stuff
All importers that open files now correctly close these resources,
whereas they previously did not and we had to hope that the garbage
collection finalizer safety net would close them for us.
Also, there were some unsafe uses of DAO Db objects that were closed
within their try block, which would be skipped if an exception occurs.
Now, a few of these Db objects are AutoCloseable, which means they can
be opened from a try-with-resources statement and always get closed.
Finally, the same applies to BatchedStatement.

ImpMetricValue had the most changes to do this properly, using a nested
class to avoid problems with field classes that needed opening/closing.
Also split up some of the networked/gzip stream handling.

Note that try-with-resources is Java 1.7+, which was already the minimum
required version for the importer.