Database: Increase size of JIRA keys
Increase from 10 to 20 characters in JIRA keys. The project key happens
to be limited to 10 characters by default by JIRA (but administrators
can increase that), and the dash and issue number add some more
characters to the key. To support importing projects with such keys,
increase the field size. Existing instances need to be upgarded too.