Update tracking: Store files in database
Keep track of the update state in the database, so that the update files
can be restored if they are lost from the data gathering workspace.

This performs the storage and updates of update file contents as well as
their modification dates. Use case is to fetch the latest versions from
the database in case the workspace files are missing or stale.