Buffered JSON: Extract and improve reader, use for issue import
The metric value buffered importer (which was not used that much since
we switched to buffered external Gzip reading) is useful in contexts
where we know how the data looks like (fancy python JSON export), since
it allows us to read the JSON objects in a buffered fashion, instead of
parsing the entire object and filling up all available memory.

Improve the buffered reader to handle more corner cases and be
applicable in other contexts that the metric value reader. Use the
buffered JSON reader for issue data, which is out of all the largest
JSON file out there.