Encryption: Convert 'encrypted' boolean into level bitmask
Instead of just noting whether a row is encrypted or not, allow the use
of encryption levels which indicate how the row was encrypted. The
levels bitmask is simply an integer, where 0 means no encryption, 1 is
encryption using project-specific salts, 2 is encryption with global
(project-independent) salt and 3 means project-specific then global salt
encryption. This can be extended later on with other masks to indicate
which parts of the row are encrypted (e.g. only personal names or also
project-identifying data, etc.)