GitLab developers: Link with VCS/global developers
Store identifiers of the developers in the merge request (notes) and
commit comments tables. This links these names to the actual developers
in most cases, but we may need to run another manual linking step.

Updated the schema to hold integers, and drop the encryption field for
now. This requires manual updates of current schemas.