Importer: New batched statement that first checks existence
The BatchedCheckStatement makes use of its own internal batching of key
values to check for existence in one massive select query, and applies
batched statements to new rows of values that need to be inserted.
Filtering is applied on the batched map of keys to remove existing rows
before they are inserted again. This provides performance improvements
over check-and-insert because the communication overhead with the
database is kept to specific batch moments.

Apply the batched check statement to commits such that a duplicate
commit with the same version ID, project ID and repo ID is not added to
the database, for example when the exported data was already added
before.