VCS: Make use of new username fields
These fields contain the login name for a review system, or a converted
variant of this name, that may more easily match JIRA user names. This
apply it to the developer name (not display name) for improved matching.
The display name is still used in the VCS developer table itself.