Metric: Use batched updates and deletes
The potential issue of a stale database with an incomplete cache can be
overcome by explicitly setting the cached value for the name to
a 'deleted' value in the cache. Thus there is no need to perform updates
immediately as long as the cache is not thrown away in the process.