Metric values: Support compact history import
For now, compact history files are detected by file name. The path
fragment may contain a date from which to import new values for the
metrics. Optimizations are made to efficiently find the date ranges that
need to be imported, using assumptions that dates and metric value
measurement ranges are sorted.