Issue: Support new import with missing fields
Instead of assuming that most/all fields are provided, and that some
fields need to be converted for '0' values, consistently check fields
that are optional and convert those fields for backward compatiblitity
if they have a '0' in them. Make some of these fields nullable if they
weren't before, and store nulls instead of things like '0'. Other fields
should always be provided and are thus marked as NOT NULL in the schema.

Provide an upgrade file for altering the database schema to work with
this. Prior upgrades were always done manually, but this makes it hard
to test, reproduce, roll back, redo or run on backups later on.