Metric: Correct missing domain name from splitting
The domain name was not correctly set after finding a matching base
name, because the domain name was appending to itself and thus remained
empty.

Prepend the domain name rather than appending, so that the domain name
parts are actually in order.

These cases can be identified as having an empty domain name but
a nonempty base name which is different from its full name.

The metric_domain_name special update task handles these situations for
upgrade purposes.