Automated schema update
This adds metadata to update files to indicate which tables/columns/keys
are being updated in the schema. This allows an update script to check
if a particular database is up to date and apply the correct scripts.
This is particularly helpful for loading in exported dumps that may not
be up to date with the schema without having to manually apply them.

Load YAML data and augment errors with filename and line context
Clean up schemas and import

- Perform update checks
- Properly handle keys and indexes
- Collect the relevant commands from the update script